STMOps.get and modify return plain A, not option
Transactions should always only use ids for data that exists - if data does not exist this is an error from which they can't recover. Therefore this failure has been moved up to the running of transactions - MapStateSTM now uses StateT with Either an Error or a value. Error is an ADT with one value at present, for IdNotFound. This means that we can catch these errors at runtime, and respond by e.g. requesting the data from a server, reporting an error, etc.
This also makes it much less fiddly to use STMOps, since we don't need to handle Option everywhere.